+++ src/ptrie.hh
@@ -6,19 +6,19 @@
 class PTrie
 {
 private:
-  std::vector<std::tuple<std::string, std::shared_ptr<PTrie>, bool, unsigned short>> v_;
+  std::vector<std::tuple<std::string, std::shared_ptr<PTrie>, bool, unsigned long>> v_;
 
 public:
   PTrie();
 
-  void insert(const std::string&, unsigned short);
+  void insert(const std::string&, unsigned long);
 
   void print(int nb_indent = 0) const;
 
-  std::vector<std::pair<std::string, unsigned short>> search(const std::string&, unsigned short) const;
+  std::vector<std::pair<std::string, unsigned long>> search(const std::string&, unsigned long) const;
 
 private:
   size_t search_prefix(const std::string&) const;
 
-  std::pair<std::string, unsigned short> search_distance0(const std::string&) const;
+  std::pair<std::string, unsigned long> search_distance0(const std::string&) const;
 };
